+++ src/main/java/service/ArchieveService.java
@@ -5,5 +5,5 @@ import domain.ArchieveList;
 import java.util.List;
 
 public interface ArchieveService {
-    List<ArchieveList> ReadList(int currentPage);
+    List<ArchieveList> ReadList(Long currentPage);
 }
